telegram-bot/requirements.txt
python-telegram-bot==21.7
aiohttp==3.9.1
pydantic==2.5.0
python-dotenv==1.0.0
python-telegram-bot==20.6

requests==2.31.0
pytz==2023.3

sqlite3 (built-in)
json (built-in)